Move promotion batch email trigger to job
